Add files via upload
My portfolio highlights some of the projects I’ve worked on, showcasing my growing knowledge and dedication to improving as a developer. I am excited to continue learning and contributing to meaningful projects as I embark on my software development journey.I am an enthusiastic and motivated beginner in software development, eager to learn and grow in the tech industry. With a foundation in [list key skills or technologies, e.g., HTML, CSS, Python, or JavaScript], I enjoy creating functional and user-friendly applications.